Update
 We apologize: this update brings Decoder API changes =(

Other changes:
 * fix spelling
 * fix some clang-tidy warnings
 * Java: move `org.brunsli` -> `dev.brunsli`
 * pull some internal API to `(enc|dec)/state.h`

+++ c/dec/ans_decode.cc
@@ -31,7 +31,7 @@ bool ANSBuildMapTable(const int counts[], int alphabet_size,
 
 }  // namespace
 
-bool ANSDecodingData::ReadFromBitStream(int alphabet_size,
+bool ANSDecodingData::ReadFromBitStream(size_t alphabet_size,
                                         BrunsliBitReader* br) {
   std::vector<int> counts(alphabet_size);
   return (ReadHistogram(ANS_LOG_TAB_SIZE, alphabet_size, &counts[0], br) &&
