+++ c/dec/ans_decode.h
@@ -26,7 +26,7 @@ typedef struct {
 struct ANSDecodingData {
   ANSDecodingData() {}
 
-  bool ReadFromBitStream(int alphabet_size, BrunsliBitReader* br);
+  bool ReadFromBitStream(size_t alphabet_size, BrunsliBitReader* br);
 
   ANSSymbolInfo map_[ANS_TAB_SIZE];
 };
@@ -37,19 +37,19 @@ class ANSDecoder {
 
   void Init(BrunsliInput* in) {
     state_ = in->GetNextWord();
-    state_ = (state_ << 16) | in->GetNextWord();
+    state_ = (state_ << 16u) | in->GetNextWord();
   }
 
   int ReadSymbol(const ANSDecodingData& code, BrunsliInput* in) {
     const uint32_t res = state_ & (ANS_TAB_SIZE - 1);
     const ANSSymbolInfo& s = code.map_[res];
     state_ = s.freq_ * (state_ >> ANS_LOG_TAB_SIZE) + s.offset_;
-    if (state_ < (1u << 16)) {
-      state_ = (state_ << 16) | in->GetNextWord();
+    if (state_ < (1u << 16u)) {
+      state_ = (state_ << 16u) | in->GetNextWord();
     }
     return s.symbol_;
   }
-  bool CheckCRC() const { return state_ == (ANS_SIGNATURE << 16); }
+  bool CheckCRC() const { return state_ == (ANS_SIGNATURE << 16u); }
 
  private:
   uint32_t state_;
